rename types.py to field_specs.py
