+++ src/ssvc/utils/field_specs.py
